+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixed "attempt to index local 'lastframe' (a nil value)" when breaking on exception
 * Removed break-on-exception when not running in Instrument Mode
 * If selected frame has a local or upval `_ENV`, evals will use it instead of the global environment
+* Removed Event Check
 ## 1.1.2
 
 * Show best-guess arguments for `__index` and `__newindex` in stack trace